requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.2.1
boto3==1.38.6
botocore==1.38.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
dropbox==12.0.2
easyocr==1.7.2
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
lazy_loader==0.4
MarkupSafe==3.0.2
motor==3.7.0
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pip==25.0
ply==3.11
propcache==0.3.1
pyasn1==0.6.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.7.1
PyJWT==2.10.1
pymongo==4.11.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.12.0
scikit-image==0.25.2
scipy==1.15.2
sendgrid==6.11.0
setuptools==75.8.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.45.3
stone==3.3.1
sympy==1.14.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
twilio==9.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
yarl==1.18.3